Fix: set player status in tournament
